Refactor into shared utilities and scan-specific file.
